+++ src/init_translations.h
@@ -10,7 +10,7 @@
 
 namespace pr {
 
-    void error_and_jacobian_translation(const Vec3f& t_ij, const Matrix3f& R_i, const Vec3f& t_i, const Vec3f& t_j, Vec3f& error, Eigen::MatrixXf& jacobians);
+    void error_and_jacobian_translation(const Vec3d& t_ij, const Matrix3d& R_i, const Vec3d& t_i, const Vec3d& t_j, Vec3d& error, Eigen::MatrixXd& jacobians);
 
     void init_translations(vector<Camera>& cameras);
